Never gonna give you up!
If a child process receives a SIGTERM/SIGINT followed by SIGHUP, it's an indication that the parent is closing down regardless of child status, but it should finish the last job it's on before dying completely. Basically, attempt a non-blocking graceful cleanup of itself rather than needing a SIGKILL to let the parent clean itself up.